requirements.txt
# 核心Web框架
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
blinker==1.8.2
click==8.1.7

# 数据库支持
SQLAlchemy==2.0.30
PyMySQL==1.1.1
alembic==1.13.1
Mako==1.3.5

# 深度学习与AI (PyTorch生态)
# 注意: 安装 torch 时请根据您的 CUDA 版本选择合适的命令，例如:
# pip install torch==2.3.1 --index-url https://download.pytorch.org/whl/cu121
torch==2.3.1
torchmetrics==1.4.0.post0
pytorch-lightning==2.3.0
transformers==4.41.2
tokenizers==0.19.1
safetensors==0.4.3
accelerate==0.31.0
diffusers==0.29.0
xformers==0.0.27
einops==0.8.0

# 科学计算与数据处理
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
scikit-learn==1.5.0
joblib==1.4.2
threadpoolctl==3.5.0

# 图像处理
Pillow==10.3.0
opencv-python==4.10.0.84
pytesseract==0.3.10
# Windows环境下推荐使用 python-magic-bin
python-magic-bin==0.4.14

# 网络请求与API集成
requests==2.32.3
urllib3==2.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
# volcengine SDK 已被移除，因项目直接使用 REST API，无需 SDK

# 工具库
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pytz>=2020.1  # 放宽版本限制以兼容 Pandas
filelock==3.15.1
psutil==5.9.8
tqdm==4.66.4
PyYAML==6.0.1
regex==2024.5.15
packaging==24.1
six==1.16.0

# 中文处理
jieba==0.42.1
zhconv==1.4.3

# 安全相关
cryptography==42.0.8
pyOpenSSL==24.1.0
cffi==1.16.0
pycryptodome==3.20.0

# 日志与调试
python-json-logger==2.0.7
retry==0.9.2

# 打包工具
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pywin32-ctypes==0.2.2
altgraph==0.17.4
